backend/requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.0.2
PyJWT==2.8.0